lua-sandbox: refactored the `sandbox` package
